+++ src/tribler/core/components/metadata_store/remote_query_community/remote_query_community.py
@@ -10,7 +10,7 @@ from ipv8.requestcache import NumberCache, RandomNumberCache, RequestCache
 from pony.orm import db_session
 from pony.orm.dbapiprovider import OperationalError
 
-from tribler.core.components.ipv8.eva_protocol import EVAProtocol, TransferResult
+from tribler.core.components.ipv8.eva.eva_protocol import EVAProtocol, TransferResult
 from tribler.core.components.ipv8.tribler_community import TriblerCommunity
 from tribler.core.components.metadata_store.db.orm_bindings.channel_metadata import LZ4_EMPTY_ARCHIVE, entries_to_chunk
 from tribler.core.components.metadata_store.db.serialization import CHANNEL_TORRENT, COLLECTION_NODE, REGULAR_TORRENT
